pyproject.toml
[project]
name = "mccube"
version = "0.0.1"
authors = [{name = "The MCCube team", email = "T.Coxon2@lboro.ac.uk"}]
description = "Markov chain cubature via JAX."
keywords = [
    "sampling",
    "probability",
    "cubature",
    "stochastic differential equations",
    "mcmc",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chex>=0.1.85",
    "diffrax>=0.4.1",
    "equinox>=0.11.1",
    "jax>=0.4.23",
    "jaxlib>=0.4.23",
    "jaxtyping>=0.2.25",
]

[project.urls]
homepage = "https://github.com/tttc3/mccube"
documentation = "https://mccube.readthedocs.io"
repository = "https://github.com/tttc3/mccube"

[project.optional-dependencies]
extras = [
    "tensorboardX>=2.6.2",
]

test = [
    "absl-py",
    "pytest",
]

dev = [
    "mccube[extras, test]",
    "jupytext",
    "pre-commit"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["mccube/*"]

[tool.pytest.ini_options]
addopts = "--jaxtyping-packages=mccube,beartype.beartype(conf=beartype.BeartypeConf(strategy=beartype.BeartypeStrategy.On))"

[tool.jupytext]
formats = "ipynb,md"

[tool.ruff]
extend-include = ["*.ipynb"]
fixable = ["I001", "F401"]
ignore = ["F722"]

[tool.ruff.isort]
combine-as-imports = true
order-by-type = false

[tool.pyright]
reportIncompatibleMethodOverride = false
include = ["mccube"]
